Renames to DPC::authorize() and DPC::execute()
