+++ dpc/src/transaction/mod.rs
@@ -17,6 +17,9 @@
 pub mod amount;
 pub use amount::*;
 
+pub mod authorization;
+pub use authorization::*;
+
 pub mod kernel;
 pub use kernel::*;
